corrected 2

+++ src/main/java/com/backend/restaurant/service/IngredientServiceImpl.java
@@ -3,6 +3,7 @@ package com.backend.restaurant.service;
 import com.backend.restaurant.model.Ingredient;
 import com.backend.restaurant.repository.IngredientRepository;
 import org.springframework.stereotype.Service;
+
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -37,11 +38,12 @@ public class IngredientServiceImpl implements IngredientService {
 
         Ingredient existing = getIngredientById(ingredient.getId());
 
-        if (existing==null) {
+        if (existing == null) {
             throw new NoSuchElementException("Ingredient with id " + ingredient.getId() + " doesn't exist.");
         }
         ingredientRepository.update(ingredient);
     }
+
     @Override
     public UUID createIngredient(String name, int quantity, BigDecimal price) {
         UUID id = UUID.randomUUID();
@@ -53,6 +55,12 @@ public class IngredientServiceImpl implements IngredientService {
 
     @Override
     public void deleteIngredient(UUID id) {
+        Ingredient existing = getIngredientById(id);
+
+        if (existing == null) {
+            throw new NoSuchElementException("Ingredient with id " + id + " doesn't exist.");
+        }
+
         ingredientRepository.delete(id);
     }
 }
